app/requirements.txt
selenium==4.10.0
webdriver-manager
python-dotenv
python-dotenv